backend_v1/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pillow
pytesseract
transformers
torch
opencv-python
numpy
scikit-learn
